requirements.txt
torch == 2.6.0
torchvision == 0.21.0
pandas == 2.2.3
numpy == 2.2.5
pillow == 11.2.1